Reworking content writing to document.
Realizing that having one global 'current' content list does
not work with storing temp page content (how to append to a page?
for the random paragraph, we can only tell up the recursion stack
that it might need to go to the page, not a random list item),
reworked that such that ODFPartialTree holds _all_ output, like
the html string in the original mistune impl.

That also nicely fixes the 'merge all textual content of a page
into *one* outline draw:frame'.