Change lpod require to odfdo
odfdo (https://github.com/jdum/odfdo) is a Python 3 library
for handling ODF documents. It is based on lpod-python which
is no longer actively maintained and still requires Python 2
to run. Since Python 2 is no longer supported by upstream and
increasingly becoming unavailable on contemporary operating
systems, it makes sense to switch to a Python 3 library.