allow H3 headings for creating slides
Many manuals allow for a three level heading TOC, like so:

* H1 --> Chapter Headings
* H2 --> Section Headings
* H3 --> Subsection Headings

So by allowing slides to be created at H3 as well, it is now possible
to maintain a single source for print, eBooks, and slides.

Hide this behind cmd line flag --lax-heading-mode, such that
accidental heading levels >2 are still causing an error by default.

Co-authored-by: githarkiratITP <harkirat.singh@itpreneurs.com>